(copilot altal generalt commit szoveg) Add Author property to Course entity
The `Course` class in the `Academy2025.Data` namespace has been updated to include a new property `Author` with a maximum length of 50 characters, marked as required.

The `ApplicationDbContextModelSnapshot` class in the `Academy2025.Migrations` namespace has been updated to include the new `Author` property for the `Course` entity, with the same constraints as defined in the `Course` class.

The `CourseRepository` class in the `Academy2025.Repositories` namespace has been updated to handle the new `Author` property when updating a course.

A new migration file `20250310210411_AddAuthorToCourse.cs` has been added to create the `Author` column in the `Courses` table in the database. This migration includes both the `Up` method to add the column and the `Down` method to remove it.

The corresponding designer file `20250310210411_AddAuthorToCourse.Designer.cs` has been generated to reflect the changes in the database schema, including the new `Author` property for the `Course` entity.